Add theme base16-tomorrow
Chris Kempson created the tomorrow theme, including the variant
'Tomorrow Night 80s', which was very popular, and already in this theme
pack. The theme evolved into the base16 project, which condenses the
number of colours down to a minimalistic 16. This i3-status theme is
using said colour scheme.

Though similar to the tomorrow-night-80s theme, this focuses on blue,
rather than green. This can be altered to the green by changing
    base0D ==> base0B
in the window and bar sections before applying.

It is also suggested to add:
    colors = true
    color_good = "#b5bd68"  # base0B
    color_bad = "#cc6666"  # base08
to the general section of your i3status.conf file to use the more mellow
green and reds from this theme.